NER_Image_Classification/requirements.txt
torch==2.0.1+cu121
torchvision==0.15.2+cu121
torchaudio==2.0.1+cu121transformers==4.30.2
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.5.0
tqdm==4.65.0
scikit-learn==1.2.2
nltk==3.8.1
opencv-python==4.7.0.72
wordcloud==1.9.2
glob2==0.7
torch.nn==2.0.1
torch.optim==2.0.1
collections==3.0
hashlib==20081119
cv2==4.7.0.72
json==2.0.9
random==1.0.3
time==1.8
datetime==4.3
itertools==8.0.0